refactor(boj22945): tighten ans to int and rework tests for verified spec bounds
- drop the long accumulator and (long) cast in Main: AC sources annotate N <= 100,000 and x_i <= 10,000, so the worst-case answer 99,998 * 10,000 fits in a signed int
- replace distinctAbilities with randomAbilities sampling WITH REPLACEMENT so the oracle exercises the duplicate-ability tie branch the spec actually allows at large N
- add cases for the spec cap, duplicate abilities, max N uniform at cap, and the distinct-only max distinct count in both directions; drop the now-unrealistic strictlyIncreasingMaxInputRequiresLongArithmetic / strictlyDecreasingMaxInputRequiresLongArithmetic
- switch @StdIo assertions to out.capturedString().trim() + isEqualTo and remove the linesOf helper to match the project's single-line BOJ output style
- expand the class javadoc with input shape, the two-pointer reduction, and the constraint sourcing; carry the trial index in the randomized oracle's .as() description

+++ boj/src/main/java/boj/boj22945/Main.java
@@ -17,11 +17,11 @@ public class Main {
         a[i] = Integer.parseInt(st.nextToken());
       }
 
-      long ans = 0;
+      int ans = 0;
       int lo = 0;
       int hi = n - 1;
       while (lo < hi) {
-        ans = Math.max(ans, (long) (hi - lo - 1) * Math.min(a[lo], a[hi]));
+        ans = Math.max(ans, (hi - lo - 1) * Math.min(a[lo], a[hi]));
         if (a[lo] < a[hi]) {
           lo++;
         } else {
